+++ packages/authhero/src/routes/universal-login/index.ts
@@ -4,6 +4,7 @@ import { DataAdapters } from "@authhero/adapter-interfaces";
 import { enterEmailRoutes } from "./enter-email";
 import { enterCodeRoutes } from "./enter-code";
 import { enterPasswordRoutes } from "./enter-password";
+import { signupRoutes } from "./signup";
 
 export interface CreateAuthParams {
   dataAdapter: DataAdapters;
@@ -18,7 +19,9 @@ export default function create() {
   const universalApp = app
     .route("/enter-email", enterEmailRoutes)
     .route("/enter-code", enterCodeRoutes)
-    .route("/enter-password", enterPasswordRoutes);
+    .route("/enter-password", enterPasswordRoutes)
+    .route("/reset-password", enterPasswordRoutes)
+    .route("/signup", signupRoutes);
 
   universalApp.doc("/u/spec", {
     openapi: "3.0.0",