+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @authhero/demo
 
+## 0.9.7
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.66.0
+
+## 0.9.6
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.65.0
+
 ## 0.9.5
 
 ### Patch Changes