+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,17 @@
 # authhero
 
+## 0.66.0
+
+### Minor Changes
+
+- Migrate signup route
+
+## 0.65.0
+
+### Minor Changes
+
+- Migarate reset password
+
 ## 0.64.0
 
 ### Minor Changes